Task/16 improve precision of recordings

See merge request hectorjsmith/csharp-performance-recorder!10

+++ PerformanceRecorder/Result/IRecordingResult.cs
@@ -20,6 +20,6 @@
 
         double Avg { get; }
 
-        void AddResult(long result);
+        void AddResult(double result);
     }
 }
